Serialize components and implement recipies
This has two parts: conversion of types to/from json and creating and
assigning them dynamically to entities. The first part is trivial and
the default implementation from nlohmann is used. The second part is a
bit more tricky and in implmented through static inline variables and
macros in ComponentRegistry.

TODO: Extend serialization to scenes and write documentation