Move shader selection from MeshRenderer to Material
Now every mesh renderer has a material instead of a shader. As a result
meshes and materials are now completely different, as they should be,
and materials now can be thought of as "prefabs for shaders", as they
contain a shader path and values for shader uniforms.

On the flipside this also means that now no material information will be
imported via assimp. This should be changed sometime in the future when
a proper importer utility will be created.