Move to a pass-based renderer implementation using RHI
This is the first step towards developing a deferred render pipeline
using RHI and not native OpenGL functions. GL implementation of RHI is
now complete and the renderer is now a namespace-level system. This
renderer should be the one to orchestrate render passes, and it should
be these passes that implement the actual rendering logic.