Remove OpenGL calls from model loader
Geometry and OpenGL buffers are now differentiated. The model loader
only loads geometry for now, but the creation of a CMeshRenderer
component triggers the creation of appropriate data buffers to be used
by the renderer. These buffers, along with all other information
required to render the model is stored in the rendering pipeline in the
form of a RenderQueue. PBRPass now renders models from this queue,
instead of from the entity registry directly. This should allow more
flexibility in the way we render the scene, and allow instancing.