Change example HDRI image
A bit of a weird problem here, the HDRI we had, after convolution, had a
few white squares at the top. Some form of an artifact that maybe was
there because the HDRI was too low-res? Anyway, switched it out for
images from sIBL, they work great.

One guy in the learnopengl comments had the same issue, didn't really
describe a solution. Will have to investigate sometime later.

https://learnopengl.com/PBR/IBL/Diffuse-irradiance