Move asset caches into the engine DLL
AssetCache::cache in AssetRegistry.h was a static variable in a header
file. What that did was create separate copies of these caches for TUs
in the user's game and TUs in the engine DLL. Caches have been moved
into the engine DLL now, and the user's TUs will use those caches only.