[engine][editor] render scene in editor window
This huge commit lays down the foundations of the interactions between
the editor and the engine. The editor creates a window, loads scenes,
loads the graphics API functions (glad) and defines the main loop. The
engine library contains functions that the editor can call to facilitate
things.

glad is built as a dll and linked to both the editor (imgui) and the
engine (renderer). The renderer just uses gl functions, it expects
gladLoad to have been called already.

glfw is only liked to the editor, because that is where the window logic
should be. All windowing logic is removed from the engine library.
Window events should be handled here and passed to ImGUI or the engine
as appropriate.

The example's main function should be removed, as the engine does that
stuff now. The example itself should now be compiled to another dll, and
called from whatever executable is handling the runtime, editor or game.